+++ packages/docs/angular-table-docs/angular-demo-module/stub.ts
@@ -1,0 +1,2 @@
+// empty file used to replace non-angular src modules so the angular compiler
+// doesn't complain. Leave this here