+++ packages/docs/angular-table-docs/src/routes/features+/row-pinning+/demos/row-pinning.service.ts
@@ -1,0 +1,8 @@
+import {Injectable, signal} from "@angular/core"
+
+@Injectable()
+export class RowPinningService {
+  readonly keepPinnedRows = signal(false)
+  readonly includeLeafRows = signal(true)
+  readonly includeParentRows = signal(true)
+}