+++ packages/docs/angular-table-docs/src/routes/features+/sorting+/demos/data.ts
@@ -1,3 +1,4 @@
+import {signal} from "@angular/core"
 import {
   type CreateQueryResult,
   injectQuery,
@@ -54,13 +55,19 @@ export const userColumns: ColumnDef<User>[] = [
 export function createUserQuery(
   ...dimensions: number[]
 ): CreateQueryResult<User[], Error> {
+  const isInitialLoad = signal(true) // [!code hide]
   return injectQuery<User[]>(() => ({
     queryFn: async () => {
-      return fetch("/get-mock-user-data", {
-        body: JSON.stringify({size: dimensions}),
+      const data = await fetch("/get-mock-user-data", {
+        body: JSON.stringify({
+          size: dimensions,
+          timestamp: isInitialLoad() ? 0 : Date.now(), // [!code hide]
+        }),
         headers: {"Content-Type": "application/json"},
         method: "POST",
       }).then((res) => res.json())
+      isInitialLoad.set(false) // [!code hide]
+      return data
     },
     queryKey: ["mock-user-data", "sorting", dimensions],
     refetchOnWindowFocus: false,